Remove test_index.py
