--- python/cudf/cudf/tests/test_index.py
@@ -1,165 +1,0 @@
-# Copyright (c) 2018-2025, NVIDIA CORPORATION.
-
-"""
-Test related to Index
-"""
-
-import operator
-
-import numpy as np
-import pandas as pd
-import pytest
-
-import cudf
-from cudf.core.index import CategoricalIndex
-from cudf.testing import assert_eq
-
-
-def test_df_set_index_from_series():
-    df = cudf.DataFrame()
-    df["a"] = list(range(10))
-    df["b"] = list(range(0, 20, 2))
-
-    # Check set_index(Series)
-    df2 = df.set_index(df["b"])
-    assert list(df2.columns) == ["a", "b"]
-    sliced_strided = df2.loc[2:6]
-    assert len(sliced_strided) == 3
-    assert list(sliced_strided.index.values) == [2, 4, 6]
-
-
-def test_df_set_index_from_name():
-    df = cudf.DataFrame()
-    df["a"] = list(range(10))
-    df["b"] = list(range(0, 20, 2))
-
-    # Check set_index(column_name)
-    df2 = df.set_index("b")
-    # 1 less column because 'b' is used as index
-    assert list(df2.columns) == ["a"]
-    sliced_strided = df2.loc[2:6]
-    assert len(sliced_strided) == 3
-    assert list(sliced_strided.index.values) == [2, 4, 6]
-
-
-def test_categorical_index():
-    pdf = pd.DataFrame()
-    pdf["a"] = [1, 2, 3]
-    pdf["index"] = pd.Categorical(["a", "b", "c"])
-    initial_df = cudf.from_pandas(pdf)
-    pdf = pdf.set_index("index")
-    gdf1 = cudf.from_pandas(pdf)
-    gdf2 = cudf.DataFrame()
-    gdf2["a"] = [1, 2, 3]
-    gdf2["index"] = pd.Categorical(["a", "b", "c"])
-    assert_eq(initial_df.index, gdf2.index)
-    gdf2 = gdf2.set_index("index")
-
-    assert isinstance(gdf1.index, CategoricalIndex)
-    assert_eq(pdf, gdf1)
-    assert_eq(pdf.index, gdf1.index)
-    assert_eq(
-        pdf.index.codes,
-        gdf1.index.codes.astype(pdf.index.codes.dtype).to_numpy(),
-    )
-
-    assert isinstance(gdf2.index, CategoricalIndex)
-    assert_eq(pdf, gdf2)
-    assert_eq(pdf.index, gdf2.index)
-    assert_eq(
-        pdf.index.codes,
-        gdf2.index.codes.astype(pdf.index.codes.dtype).to_numpy(),
-    )
-
-
-def test_set_index_as_property():
-    cdf = cudf.DataFrame()
-    col1 = np.arange(10)
-    col2 = np.arange(0, 20, 2)
-    cdf["a"] = col1
-    cdf["b"] = col2
-
-    # Check set_index(Series)
-    cdf.index = cdf["b"]
-
-    assert_eq(cdf.index.to_numpy(), col2)
-
-    with pytest.raises(ValueError):
-        cdf.index = [list(range(10))]
-
-    idx = pd.Index(np.arange(0, 1000, 100))
-    cdf.index = idx
-    assert_eq(cdf.index.to_pandas(), idx)
-
-    df = cdf.to_pandas()
-    assert_eq(df.index, idx)
-
-    head = cdf.head().to_pandas()
-    assert_eq(head.index, idx[:5])
-
-
-@pytest.mark.parametrize(
-    "n",
-    [-10, -5, -2, 0, 1, 0, 2, 5, 10],
-)
-def test_empty_df_head_tail_index(n):
-    df = cudf.DataFrame()
-    pdf = pd.DataFrame()
-    assert_eq(df.head(n).index.values, pdf.head(n).index.values)
-    assert_eq(df.tail(n).index.values, pdf.tail(n).index.values)
-
-    df = cudf.DataFrame({"a": [11, 2, 33, 44, 55]})
-    pdf = pd.DataFrame({"a": [11, 2, 33, 44, 55]})
-    assert_eq(df.head(n).index.values, pdf.head(n).index.values)
-    assert_eq(df.tail(n).index.values, pdf.tail(n).index.values)
-
-    df = cudf.DataFrame(index=[1, 2, 3])
-    pdf = pd.DataFrame(index=[1, 2, 3])
-    assert_eq(df.head(n).index.values, pdf.head(n).index.values)
-    assert_eq(df.tail(n).index.values, pdf.tail(n).index.values)
-
-
-@pytest.mark.parametrize(
-    "objs",
-    [
-        [pd.RangeIndex(0, 10), pd.RangeIndex(10, 20)],
-        [pd.RangeIndex(10, 20), pd.RangeIndex(22, 40), pd.RangeIndex(50, 60)],
-        [pd.RangeIndex(10, 20, 2), pd.RangeIndex(20, 40, 2)],
-    ],
-)
-def test_range_index_concat(objs):
-    cudf_objs = [cudf.from_pandas(obj) for obj in objs]
-
-    actual = cudf.concat(cudf_objs)
-
-    expected = objs[0]
-    for obj in objs[1:]:
-        expected = expected.append(obj)
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "op", [operator.add, operator.sub, operator.mul, operator.truediv]
-)
-def test_rangeindex_binop_diff_names_none(op):
-    idx1 = cudf.RangeIndex(10, 13, name="foo")
-    idx2 = cudf.RangeIndex(13, 16, name="bar")
-    result = op(idx1, idx2)
-    expected = op(idx1.to_pandas(), idx2.to_pandas())
-    assert_eq(result, expected)
-    assert result.name is None
-
-
-@pytest.mark.parametrize(
-    "index",
-    [
-        lambda: cudf.Index([1]),
-        lambda: cudf.RangeIndex(1),
-        lambda: cudf.MultiIndex(levels=[[0]], codes=[[0]]),
-    ],
-)
-def test_index_assignment_no_shallow_copy(index):
-    index = index()
-    df = cudf.DataFrame(range(1))
-    df.index = index
-    assert df.index is index
